chatgpt/server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
blobfile==2.0.2
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
filelock==3.12.4
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-RESTful==0.3.9
flask-restx==0.5.1
#Flask-RESTPlus==0.13.0
# flasgger==0.9.7.1
flower==1.1.0
frozenlist==1.4.0
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
humanize==4.8.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
limits==3.6.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openai==1.30.3
openai[datalib]
OpenAIAuth==0.3.2
ordered-set==4.1.0
packaging==23.1
peewee==3.16.0
peewee-migrate==1.6.6
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.5
pycryptodomex==3.19.0
Pygments==2.16.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
regex==2023.8.8
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.5.3
six==1.16.0
supervisor==4.2.2
tenacity==8.2.2
tiktoken==0.3.0
tokenizers==0.13.3
tornado==6.3.3
tqdm==4.66.1
tree-sitter==0.20.2
tree-sitter-languages==1.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.15  # 1.26.16版本有毒，禁止使用
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
wtf-peewee==3.0.4
WTForms==3.0.1
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
pydantic==2.5.3
json_repair==0.8.0
python-redis-lock==4.0.0
pyautogen==0.2.27
uvloop==0.19.0
python-socketio==5.11.2
